refactor: split VM scanning utils into framework and test files
Separate framework helpers (config loading, env parsing, SSH key
generation, virtctl discovery) from their associated test functions
for better file organization.

Partially AI-generated.

Made-with: Cursor
